--- integration_tests/test_upgrade_testnet.py
@@ -1,32 +1,0 @@
-import subprocess
-from pathlib import Path
-
-import pytest
-
-from .network import Cronos, setup_custom_cronos
-from .upgrade_utils import exec, post_init
-
-
-@pytest.fixture(scope="module")
-def custom_cronos(tmp_path_factory):
-    path = tmp_path_factory.mktemp("upgrade")
-    cmd = [
-        "nix-build",
-        Path(__file__).parent / "configs/upgrade-testnet-test-package.nix",
-        "-o",
-        path / "upgrades",
-    ]
-    print(*cmd)
-    subprocess.run(cmd, check=True)
-    # init with genesis binary
-    yield from setup_custom_cronos(
-        path,
-        26200,
-        Path(__file__).parent / "configs/cosmovisor_testnet.jsonnet",
-        post_init=post_init,
-        chain_binary=str(path / "upgrades/genesis/bin/cronosd"),
-    )
-
-
-def test_cosmovisor_upgrade(custom_cronos: Cronos, tmp_path_factory):
-    exec(custom_cronos, tmp_path_factory)
